blofin-bridge: float sizing (fixes precision loss on fractional-SOL sizes)
Previous code returned int(contracts) which dropped fractional SOL. For
$100/10x at $84.7, old code returned 11 (= $931 notional), new code
returns 11.80 (= $999 notional). 7% more size, matches target $100
margin almost exactly.

- sizing.py: return float, new _floor_to_lot helper with IEEE754 noise guard
- schema.sql: initial_size/current_size now REAL
- state.py PositionRow + create_position + record_tp_fill: float types
- policies/base.py Position + SLOrder: float types
- blofin_client.py place_market_entry/close_position_market: contracts float
- conftest sol_instrument: real SOL-USDT lotSize=0.01 minSize=0.01
- tests updated for float expectations

+++ scripts/blofin-bridge/src/blofin_bridge/blofin_client.py
@@ -77,7 +77,7 @@ class BloFinClient:
         )
 
     def place_market_entry(
-        self, *, inst_id: str, side: str, contracts: int,
+        self, *, inst_id: str, side: str, contracts: float,
         safety_sl_trigger: float,
     ) -> dict[str, Any]:
         """Market entry with an attached safety SL (OCO-style).
@@ -188,7 +188,7 @@ class BloFinClient:
         return cancelled
 
     def close_position_market(
-        self, *, inst_id: str, side: str, contracts: int,
+        self, *, inst_id: str, side: str, contracts: float,
     ) -> dict[str, Any]:
         """Reduce-only market order to close N contracts.
